player speed upgrade working

+++ Game-Mod-Q4/game/Player.h
@@ -289,7 +289,7 @@ public:
 	int						GL_UPGRADE_FIRE_RATE;
 	float					GL_UPGRADE_SPREAD;
 	int						GL_UPGRADE_PROJ_COUNT;
-	int						PLAYER_SPEED_UPGRADE;
+	float					PLAYER_SPEED_UPGRADE;
 
 	void					doTheThing(int amountToAdd);
 	void					doTheReverseThing(int amountToSubtract);
